rename normalize props proxy example folder
