--- examples/TEMP-local-state-with-fix-fixture-props-proxy/cosmos.config.js
@@ -1,4 +1,0 @@
-export default {
-  componentPaths: ['components'],
-  proxies: ['./fix-fixture-props-proxy'],
-};
