+++ examples/local-state-with-normalize-props-proxy/components/Counter/Counter.jsx
