+++ examples/local-state-with-normalize-props-proxy/components/Counter/__fixtures__/default.js
