+++ examples/local-state-with-normalize-props-proxy/components/Counter/__fixtures__/fiveClicks.js
