+++ examples/local-state-with-normalize-props-proxy/components/CounterList/CounterList.jsx
