+++ examples/local-state-with-normalize-props-proxy/components/CounterList/__fixtures__/oneTwoThree.js
