+++ examples/local-state-with-normalize-props-proxy/cosmos.config.js
@@ -1,0 +1,5 @@
+export default {
+  componentPaths: ['components'],
+  // TODO: use package for normalize proxy
+  proxies: ['../../packages/react-cosmos-normalize-props-proxy/src']
+};
